// src-ui/core/ipc/generated/domain/HttpClientConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MessagePostFormat } from "./MessagePostFormat";

export type HttpClientConfig = { url: string, reportSelfMessage: boolean, timeoutMs?: number, enable: boolean, name: string, messagePostFormat: MessagePostFormat, token: string, debug: boolean, };